pyproject.toml
[project]
name = "ha-timescaledb-recorder"
version = "0.1.0"
requires-python = ">=3.14.2"

[tool.pytest]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "asyncpg==0.31.0",
    "psycopg[binary]==3.3.3",
    "pytest>=9.0",
    "pytest-asyncio>=0.24",
    "pytest-homeassistant-custom-component>=0.13.320",
]
